well folks, we got a chicken on the screen
that was the goal, it looks like poop, but that was the goal. next steps
are to make the chicken look good :)

quite a lot of changes went into this commit as you can see. i had to
revamp the shader and program system at least a couple of times. i'm
probably like 50% confident that the system in place is future-proof and
very well organized for future program and shader creation. it may need
more tweaking as development goes on.

still left to do are the following, and probably more but i'm just not
thinking right now:
- sprite scaling
- double check camera aspect ratio
- is glfw responding to window resizing?
- alpha on textures (the chicken has black around it)

sprite animation, etc are a whole other task. right now i think it's
best to stick to rendering static images with the correct size and shape
on the screen.

hooray! this is a *HUGE* step forward!